@@ -1,16 +1,16 @@
 name: license_plate_number
 description: Flutter widget for inputting license plate number. Build-in one light style and one dark style, and the style can be changed.
-version: 1.1.0
+version: 2.0.0
 homepage: https://github.com/chenglei1986/license_plate_number
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  dotted_border: ^1.0.5
+  dotted_border: ^2.0.0-nullsafety.0
 
 dev_dependencies:
   flutter_test:
