@@ -7,16 +7,15 @@ class SinglePlateNumberField {
     this.text, {
     this.textStyle,
     this.decoration,
-  })  : assert(text != null, 'Plate number must be a non-null String.'),
-        assert(text.length < 2,
+  })  : assert(text.length < 2,
             'SinglePlateNumberField can only accept single character as its content');
 
   /// 车牌号单个字符
   final String text;
 
   /// 字体样式
-  final TextStyle textStyle;
+  final TextStyle? textStyle;
 
   /// 边框及背景
-  final Decoration decoration;
+  final Decoration? decoration;
 }
